+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## New Features
 
+-   `reconstruct_qsm()`: New function to reconstruct a QSM and all of its variables from the minimum amount of required data. This enables the reconstruction and analysis of generic QSMs from external software or even manual measurements. The only required variables are the cylinder id, parent id, 3d information (e.g. a combination of start and end points, or start, axis, and length), and the radius.
 -   `export_mesh()`: add [GroIMP](https://grogra.de/) as a QSM export format with `format = "groimp"` (issue #21).
 
 # rTwig 1.4.0